cp-78: * rename folder and file from backward-buton to back-button
