+++ frontend/src/libs/components/back-button/styles.module.scss
